Lab6/requirements.txt
ocrmypdf==17.3.0
pdfplumber==0.11.9
pytesseract==0.3.13
mysql-connector-python==9.6.0
python-dotenv==1.2.1
requests==2.32.3
beautifulsoup4==4.14.3
flask==3.1.3
